Added useFormattedCode hook
